refactor: Remplace Input component by UI-kit Input (#556)
* refactor: Update with Input component in BalanceInput

* refactor: Update with Input component in TicketInput

* refactor: Update with Input component in TokenInput

* refactor: Remove Input component

* refactor: Move BalanceInput component and update imports

* refactor: Change onChange prop to non-optional
