+++ src/components/BalanceInput/BalanceInput.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
 import styled from 'styled-components'
-import { Button, Box, BoxProps, Flex, Text } from '@pancakeswap-libs/uikit'
+import { Button, Box, BoxProps, Flex, Input, Text } from '@pancakeswap-libs/uikit'
 import useI18n from 'hooks/useI18n'
-import Input from './Input'
 
-interface BalanceInputProps extends BoxProps {
+export interface BalanceInputProps extends BoxProps {
   title: string
   max: number
   symbol: string
@@ -19,7 +18,6 @@ const StyledBalanceInput = styled(Box)`
   border-radius: ${({ theme }) => theme.radii.default};
   padding: 8px 16px;
 `
-
 const BalanceInput: React.FC<BalanceInputProps> = ({ title, max, symbol, onChange, onSelectMax, value, ...props }) => {
   const TranslateString = useI18n()
   const maxDisplay = max.toFixed(6)
@@ -30,21 +28,17 @@ const BalanceInput: React.FC<BalanceInputProps> = ({ title, max, symbol, onChang
         <Text fontSize="14px">{title}</Text>
         <Text fontSize="14px">{TranslateString(999, `Balance: ${maxDisplay}`, { num: maxDisplay })}</Text>
       </Flex>
-      <Input
-        endAdornment={
-          <Flex alignItems="center">
-            {onSelectMax && (
-              <Button scale="sm" onClick={onSelectMax} mr="8px">
-                {TranslateString(452, 'Max')}
-              </Button>
-            )}
-            <Text>{symbol}</Text>
-          </Flex>
-        }
-        onChange={onChange}
-        placeholder="0"
-        value={value}
-      />
+      <Flex alignItems="center">
+        <Input onChange={onChange} placeholder="0" value={value} />
+        <Flex alignItems="center">
+          {onSelectMax && (
+            <Button scale="sm" onClick={onSelectMax} mr="8px">
+              {TranslateString(452, 'Max')}
+            </Button>
+          )}
+          <Text>{symbol}</Text>
+        </Flex>
+      </Flex>
     </StyledBalanceInput>
   )
 }
