--- src/components/Input/index.ts
@@ -1,2 +1,0 @@
-export { default } from './Input'
-export type { InputProps } from './Input'
