+++ src/components/TicketInput/TicketInput.tsx
@@ -1,14 +1,15 @@
 import React from 'react'
 import styled from 'styled-components'
-import { Button } from '@pancakeswap-libs/uikit'
+import { Button, Flex, Input, InputProps } from '@pancakeswap-libs/uikit'
 import useI18n from '../../hooks/useI18n'
-import Input, { InputProps } from '../Input'
 
 interface TokenInputProps extends InputProps {
   max: number | string
   symbol: string
   availableSymbol: string
+  value: string
   onSelectMax?: () => void
+  onChange: (evt: React.FormEvent<HTMLInputElement>) => void
 }
 
 const TicketInput: React.FC<TokenInputProps> = ({ max, symbol, availableSymbol, onChange, onSelectMax, value }) => {
@@ -16,22 +17,18 @@ const TicketInput: React.FC<TokenInputProps> = ({ max, symbol, availableSymbol, 
 
   return (
     <StyledTokenInput>
-      <Input
-        endAdornment={
-          <StyledTokenAdornmentWrapper>
-            <StyledTokenSymbol>{symbol}</StyledTokenSymbol>
-            <StyledSpacer />
-            <div>
-              <Button scale="sm" onClick={onSelectMax}>
-                {TranslateString(452, 'Max')}
-              </Button>
-            </div>
-          </StyledTokenAdornmentWrapper>
-        }
-        onChange={onChange}
-        placeholder="0"
-        value={value}
-      />
+      <Flex alignItems="center">
+        <Input onChange={onChange} placeholder="0" value={value} />
+        <StyledTokenAdornmentWrapper>
+          <StyledTokenSymbol>{symbol}</StyledTokenSymbol>
+          <StyledSpacer />
+          <div>
+            <Button size="sm" onClick={onSelectMax}>
+              {TranslateString(452, 'Max')}
+            </Button>
+          </div>
+        </StyledTokenAdornmentWrapper>
+      </Flex>
       <StyledMaxText>{TranslateString(454, `${max.toLocaleString()} ${availableSymbol} Available`)}</StyledMaxText>
     </StyledTokenInput>
   )
