+++ src/views/Ifos/components/IfoCard/ContributeModal.tsx
@@ -3,7 +3,7 @@ import { useWeb3React } from '@web3-react/core'
 import BigNumber from 'bignumber.js'
 import { ethers } from 'ethers'
 import { Modal, LinkExternal, Box } from '@pancakeswap-libs/uikit'
-import BalanceInput from 'components/Input/BalanceInput'
+import BalanceInput from 'components/BalanceInput'
 import useTokenBalance from 'hooks/useTokenBalance'
 import { getBalanceNumber } from 'utils/formatBalance'
 import useI18n from 'hooks/useI18n'
